More progress
More progress made on both backend and front end. Access and refresh tokens on the backend are looking good, and now they're recorded in the database to allow token revocation. On the front end we've properly handled error showing for the signup and login form it seems like, and we're successfully logging in and setting the global state for the AuthProvider. However, it still needs some work as we need to make it persistent, so trying to refresh when they've refreshed and refreshing the token when it's expired, which just requires us knowing when a request failed because of the expiration. We may need to learn axios a bit though to understand the Dave Gray tutorial