Progress on front and backend
In our backend we've configured our server to only accept certain origins and we've successfully been able to send and receive cookies. We've added roles to our users. Also we've cleaned up some error handling, so that sometimes we'll just send back json data as the error immediately instead of making and throwing an error down our error pipeline. As a result, only unexpected errors will be thrown down our error pipeline. Added middleware that allows us to verify a user's role before accessing an api route, and we'll make use of that in later when we start fleshing out our api routes. On the front end we've replaced fetch with axios. As well as this we've been successful in logging in, logging out, and persisting a user's authentication state. Now we've been mainly working on our BrowsePage and doing responsive design there.