Major progress on browse page, dashboard, and profile page
Browse page now has quality sticky header, sidebar, and scrollable content. Added a simplified dark mode throughout the front end application. Is responsive and has youtube-like cards to show off the articles, but right now we're just using mock articles to see it. Account menu now displays the logged in user's profile picture, and it's able to truncate user info if it's too long. As well as this it is able to successfully navigate the user to the Dashboard. The dashboard also has that sticky header, sidebar design, with a scrollable main content, which is good. The current tab the user is on will be highlighted as well. The profile page has a modal form that lets the user change their avatar. We also have forms that let the user change their other info, but those haven't been fleshed out yet. Also we finally migrated from using a simple state to having a reducer manage our auth state. On the backend we set up multer to store user avatars/profile pictures on disk, and also track that in the database for users. We also improved the User model, as with the .method.toJSON() function, we fully control what is returned when a user is sent as json. With that, we successfully retained a user model's virtual properties and we're able to filter out what fields we want and don't want being sent back in json. All of this json management done in one function.