+ Front end: - Fixed rendering for error messages on ManageCategoriesPage, ManagePostsPage, and ManageTagsPage. - Updated EmployeeGrid to display a better error message when lastLogin is undefined for a row.
+ Back end:
- Implemented Redis Caching with a local redis server. At the end we'll go to a free cloud server. Now we're able to efficiently cache information such as tags, categories, and employees.

+ Upcoming:
For our last changes, I'm planning to learn about and at least add actual logging to our backend-express server since placing console.logs isn't that good. As well as this, I plan to migrate the front-end to TypeScript as well. After we'll do Redis cloud, and then deployment to render.