Still incomplete, but still making a little more progress on frontend's typescript
One thing I'll note is that this TypeScript Migration will force the project to adopt a proper MUI color and theme setup, so yeah after updating the typescript and making sure everything works, we'll need to create a new theme setup. Fortunately we have a model for that, but of course having ag ood setup is the hard part.